requirements.txt
numpy
nltk
huggingface_hub
pymupdf
clean-text
python-dotenv